+++ ios/Classes/InAppWebView.swift
@@ -745,6 +745,63 @@ public class InAppWebView: WKWebView, UIScrollViewDelegate, WKUIDelegate, WKNavi
         }
     }
     
+    public func webView(_ webView: WKWebView, runJavaScriptAlertPanelWithMessage message: String,
+                 initiatedByFrame frame: WKFrameInfo, completionHandler: @escaping () -> Void) {
+
+        let alertController = UIAlertController(title: message, message: nil,
+                                                preferredStyle: UIAlertController.Style.alert);
+
+        alertController.addAction(UIAlertAction(title: "OK", style: UIAlertAction.Style.default) {
+            _ in completionHandler()}
+        );
+        
+        let presentingViewController = ((IABController != nil) ? IABController! : window!.rootViewController!)
+        presentingViewController.present(alertController, animated: true, completion: {})
+    }
+    
+    public func webView(_ webView: WKWebView, runJavaScriptConfirmPanelWithMessage message: String, initiatedByFrame frame: WKFrameInfo,
+                 completionHandler: @escaping (Bool) -> Void) {
+
+        let alertController = UIAlertController(title: nil, message: message, preferredStyle: .alert)
+
+        alertController.addAction(UIAlertAction(title: "OK", style: .default, handler: { (action) in
+            completionHandler(true)
+        }))
+
+        alertController.addAction(UIAlertAction(title: "Cancel", style: .cancel, handler: { (action) in
+            completionHandler(false)
+        }))
+
+        let presentingViewController = ((IABController != nil) ? IABController! : window!.rootViewController!)
+        presentingViewController.present(alertController, animated: true, completion: nil)
+    }
+
+
+    public func webView(_ webView: WKWebView, runJavaScriptTextInputPanelWithPrompt prompt: String, defaultText: String?, initiatedByFrame frame: WKFrameInfo,
+                 completionHandler: @escaping (String?) -> Void) {
+
+        let alertController = UIAlertController(title: nil, message: prompt, preferredStyle: .alert)
+
+        alertController.addTextField { (textField) in
+            textField.text = defaultText
+        }
+
+        alertController.addAction(UIAlertAction(title: "OK", style: .default, handler: { (action) in
+            if let text = alertController.textFields?.first?.text {
+                completionHandler(text)
+            } else {
+                completionHandler(defaultText)
+            }
+        }))
+
+        alertController.addAction(UIAlertAction(title: "Cancel", style: .cancel, handler: { (action) in
+            completionHandler(nil)
+        }))
+
+        let presentingViewController = ((IABController != nil) ? IABController! : window!.rootViewController!)
+        presentingViewController.present(alertController, animated: true, completion: nil)
+    }
+    
     public func scrollViewDidScroll(_ scrollView: UIScrollView) {
         if navigationDelegate != nil {
             let x = Int(scrollView.contentOffset.x / scrollView.contentScaleFactor)